initial pass at updating packages
