--- service/.eslintignore
@@ -1,3 +1,0 @@
-**/node_modules/**
-**/dist**
-.eslintrc.js
