+++ service/eslint.config.js
@@ -1,0 +1,71 @@
+const { defineConfig, globalIgnores } = require('eslint/config');
+
+const tsParser = require('@typescript-eslint/parser');
+const typescriptEslintEslintPlugin = require('@typescript-eslint/eslint-plugin');
+const prettier = require('eslint-plugin-prettier');
+const globals = require('globals');
+const js = require('@eslint/js');
+
+const { FlatCompat } = require('@eslint/eslintrc');
+
+const compat = new FlatCompat({
+  baseDirectory: __dirname,
+  recommendedConfig: js.configs.recommended,
+  allConfig: js.configs.all,
+});
+const env = (prod, dev) => (process.env.NODE_ENV === 'production' ? prod : dev);
+
+module.exports = defineConfig([
+  {
+    languageOptions: {
+      parser: tsParser,
+      sourceType: 'module',
+
+      parserOptions: {
+        project: 'tsconfig.json',
+      },
+
+      globals: {
+        ...globals.node,
+        ...globals.jest,
+      },
+    },
+
+    plugins: {
+      '@typescript-eslint': typescriptEslintEslintPlugin,
+      prettier,
+    },
+
+    extends: compat.extends(
+      'plugin:@typescript-eslint/eslint-recommended',
+      'plugin:@typescript-eslint/recommended',
+      'plugin:prettier/recommended',
+      'prettier'
+    ),
+
+    rules: {
+      quotes: ['error', 'single'],
+      'no-console': env(1, 0),
+      'no-debugger': env(1, 0),
+      '@typescript-eslint/interface-name-prefix': 'off',
+      '@typescript-eslint/explicit-function-return-type': 'off',
+      '@typescript-eslint/explicit-module-boundary-types': 'off',
+      '@typescript-eslint/no-explicit-any': 'off',
+
+      '@typescript-eslint/no-unused-vars': [
+        env(2, 1),
+        {
+          argsIgnorePattern: '^_',
+        },
+      ],
+
+      'no-multiple-empty-lines': 'error',
+    },
+  },
+  globalIgnores([
+    '**/node_modules/**/*',
+    '**/dist**',
+    '**/.eslintrc.js',
+    'src/migrations',
+  ]),
+]);
