+++ service/package.json
@@ -1,11 +1,12 @@
 {
   "name": "alkemio-matrix-adapter",
-  "version": "0.4.6",
+  "version": "0.5.0",
   "description": "Alkemio Matrix Adapter service",
   "author": "Alkemio Foundation",
   "private": false,
   "license": "EUPL-1.2",
   "main": "src/main.ts",
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/alkem-io/matrix-adapter.git"
@@ -35,51 +36,49 @@
   },
   "dependencies": {
     "@alkemio/matrix-adapter-lib": "^0.4.1",
-    "@nestjs/axios": "^2.0.0",
-    "@nestjs/common": "^9.4.0",
-    "@nestjs/config": "^2.3.1",
-    "@nestjs/core": "^9.4.0",
-    "@nestjs/microservices": "^9.0.0",
-    "@nestjs/platform-express": "^9.4.0",
+    "@nestjs/axios": "^3.1.3",
+    "@nestjs/common": "^10.4.19",
+    "@nestjs/config": "^3.2.3",
+    "@nestjs/core": "^10.3.10",
+    "@nestjs/microservices": "^10.3.10",
     "amqp-connection-manager": "^3.7.0",
     "amqplib": "^0.8.0",
     "cross-env": "^7.0.3",
     "crypto-js": "^4.1.1",
-    "dotenv": "^10.0.0",
-    "matrix-js-sdk": "^32.4.0",
-    "module-alias": "^2.2.2",
-    "nest-winston": "^1.9.1",
-    "winston": "^3.3.3",
-    "winston-elasticsearch": "^0.15.8",
-    "yaml": "^2.0.0-7"
+    "dotenv": "^16.5.0",
+    "matrix-js-sdk": "^35.0.0",
+    "module-alias": "^2.2.3",
+    "nest-winston": "^1.10.2",
+    "winston": "^3.13.1",
+    "winston-elasticsearch": "^0.19.0",
+    "yaml": "^2.4.5"
   },
   "devDependencies": {
-    "@nestjs/cli": "^9.4.0",
-    "@nestjs/testing": "^9.4.0",
-    "@types/amqplib": "^0.8.2",
-    "@types/crypto-js": "^4.0.2",
-    "@types/express": "^4.17.13",
-    "@types/jest": "^26.0.24",
-    "@types/module-alias": "^2.0.1",
-    "@types/node": "^16.4.6",
+    "@nestjs/cli": "^10.4.2",
+    "@types/amqplib": "^0.10.5",
+    "@types/crypto-js": "^4.2.2",
+    "@types/express": "^5.0.3",
+    "@types/jest": "^30.0.0",
+    "@types/module-alias": "^2.0.4",
+    "@types/node": "^20.14.10",
     "@types/yaml": "^1.9.7",
-    "@typescript-eslint/eslint-plugin": "^5.59.1",
-    "@typescript-eslint/parser": "^5.59.1",
+    "@typescript-eslint/eslint-plugin": "^8.32.1",
+    "@typescript-eslint/parser": "^8.32.1",
     "cd": "^0.3.3",
     "coveralls": "^3.1.1",
-    "eslint": "^8.39.0",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-prettier": "^3.4.0",
-    "husky": "^4.3.8",
-    "jest": "^27.0.6",
-    "lint-staged": "^13.2.2",
-    "prettier": "^2.3.2",
-    "rimraf": "^3.0.2",
-    "ts-jest": "^27.0.4",
-    "ts-node": "^10.8.1",
-    "tsconfig-paths": "^4.1.1",
-    "typescript": "^4.9.4"
+    "eslint": "^9.26.0",
+    "eslint-config-prettier": "^10.1.5",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-prettier": "^5.4.0",
+    "husky": "^9.1.7",
+    "jest": "^29.7.0",
+    "lint-staged": "^16.1.2",
+    "prettier": "^3.5.3",
+    "rimraf": "^6.0.1",
+    "ts-jest": "^29.2.5",
+    "tsx": "^4.19.4",
+    "tsconfig-paths": "^4.2.0",
+    "typescript": "5.3.3"
   },
   "_moduleAliases": {
     "@src": "./dist"
